+++ sdk/resourcemanager/containerservice/armcontainerservice/models.go
@@ -150,6 +150,13 @@ type AgentPoolUpgradeSettings struct {
 	NodeSoakDurationInMinutes *int32
 }
 
+// AgentPoolWindowsProfile - The Windows agent pool's specific profile.
+type AgentPoolWindowsProfile struct {
+	// The default value is false. Outbound NAT can only be disabled if the cluster outboundType is NAT Gateway and the Windows
+	// agent pool does not have node public IP enabled.
+	DisableOutboundNat *bool
+}
+
 // AzureKeyVaultKms - Azure Key Vault key management service settings for the security profile.
 type AzureKeyVaultKms struct {
 	// Whether to enable Azure Key Vault key management service. The default is false.
@@ -816,6 +823,9 @@ type ManagedClusterAgentPoolProfile struct {
 	// /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/virtualNetworks/{virtualNetworkName}/subnets/{subnetName}
 	VnetSubnetID *string
 
+	// The Windows agent pool's specific profile.
+	WindowsProfile *AgentPoolWindowsProfile
+
 	// Determines the type of workload a node can run.
 	WorkloadRuntime *WorkloadRuntime
 
@@ -978,6 +988,9 @@ type ManagedClusterAgentPoolProfileProperties struct {
 	// /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/virtualNetworks/{virtualNetworkName}/subnets/{subnetName}
 	VnetSubnetID *string
 
+	// The Windows agent pool's specific profile.
+	WindowsProfile *AgentPoolWindowsProfile
+
 	// Determines the type of workload a node can run.
 	WorkloadRuntime *WorkloadRuntime
 
@@ -1071,6 +1084,31 @@ type ManagedClusterIdentity struct {
 
 	// READ-ONLY; The tenant id of the system assigned identity which is used by master components.
 	TenantID *string
+}
+
+// ManagedClusterIngressProfile - Ingress profile for the container service cluster.
+type ManagedClusterIngressProfile struct {
+	// App Routing settings for the ingress profile. You can find an overview and onboarding guide for this feature at
+	// https://learn.microsoft.com/en-us/azure/aks/app-routing?tabs=default%2Cdeploy-app-default.
+	WebAppRouting *ManagedClusterIngressProfileWebAppRouting
+}
+
+// ManagedClusterIngressProfileWebAppRouting - Application Routing add-on settings for the ingress profile.
+type ManagedClusterIngressProfileWebAppRouting struct {
+	// Resource IDs of the DNS zones to be associated with the Application Routing add-on. Used only when Application Routing
+	// add-on is enabled. Public and private DNS zones can be in different resource
+	// groups, but all public DNS zones must be in the same resource group and all private DNS zones must be in the same resource
+	// group.
+	DNSZoneResourceIDs []*string
+
+	// Whether to enable the Application Routing add-on.
+	Enabled *bool
+
+	// READ-ONLY; Managed identity of the Application Routing add-on. This is the identity that should be granted permissions,
+	// for example, to manage the associated Azure DNS resource and get certificates from Azure
+	// Key Vault. See this overview of the add-on [https://learn.microsoft.com/en-us/azure/aks/web-app-routing?tabs=with-osm]
+	// for more instructions.
+	Identity *UserAssignedIdentity
 }
 
 // ManagedClusterListResult - The response from the List Managed Clusters operation.
@@ -1316,6 +1354,9 @@ type ManagedClusterProperties struct {
 
 	// Identities associated with the cluster.
 	IdentityProfile map[string]*UserAssignedIdentity
+
+	// Ingress profile for the managed cluster.
+	IngressProfile *ManagedClusterIngressProfile
 
 	// Both patch version (e.g. 1.20.13) and (e.g. 1.20) are supported. When is specified, the latest supported GA patch version
 	// is chosen automatically. Updating the cluster with the same once it has been